Refactor helpers to be a folder with different files for different purposes
