+++ lib/helpers/helpers.rb
@@ -1,0 +1,6 @@
+def parse_response(response)
+  is_data_ok = false if %w[n N].include? response
+  is_data_ok = true if %w[y Y].include? response
+
+  is_data_ok
+end
