+++ lib/helpers/object_creation.rb
@@ -1,31 +1,8 @@
-require_relative 'classes/movie'
-require_relative 'classes/genre'
-require_relative 'classes/label'
-require_relative 'classes/author'
-require 'date'
-
-def list_all(arr, label)
-  puts "List of all #{label}"
-  print 'title | author | genre | source'
-  arr.each_with_index do |item, index|
-    print (index + 1).to_s
-    print "#{item.label.title} | "
-    print "#{item.author.first_name} "
-    print "#{item.author.last_name} | "
-    print "#{item.genre.name} | "
-    print "#{item.source.name} \n"
-  end
-end
-
-def add_element(element)
-  puts "Here is the method to add element #{element}"
-end
-
-def exit
-  puts 'Thanks for using this App'
-def generate_id
-  rand(1..1000)
-end
+require_relative '../classes/movie'
+require_relative '../classes/genre'
+require_relative '../classes/label'
+require_relative '../classes/author'
+require_relative 'helpers'
 
 def create_movie
   id = generate_id
@@ -106,11 +83,4 @@ def create_author
   las_name = gets.chomp
 
   Author.new(id, name, las_name)
-end
-
-def parse_response(response)
-  is_data_ok = false if %w[n N].include? response
-  is_data_ok = true if %w[y Y].include? response
-
-  is_data_ok
 end
